Renamed namespace to WearHouse_WebApp
