+++ WearHouse/WearHouse_WebApp/Data/ApplicationDbContext.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Prototype02.Data
+namespace WearHouse_WebApp.Data
 {
     public class ApplicationDbContext : IdentityDbContext
     {
