+++ WearHouse/WearHouse_WebApp/Models/ErrorViewModel.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Prototype02.Models
+namespace WearHouse_WebApp.Models
 {
     public class ErrorViewModel
     {
