+++ WearHouse/WearHouse_WebApp/Startup.cs
@@ -7,13 +7,13 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Prototype02.Data;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using WearHouse_WebApp.Data;
 
-namespace Prototype02
+namespace WearHouse_WebApp
 {
     public class Startup
     {
