+++ WearHouse/WearHouse_WebApp/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
-﻿@using Prototype02
-@using Prototype02.Models
+﻿@using WearHouse_WebApp
+@using WearHouse_WebApp.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
